:sparkles: add computed knobbing information
This commit adds a `KnobInfo` structure to the setting property. Both
reading and setting properties now extend the `DeviceProperty` interface
which forces them to have the same base set of fields. For settings, we
add the optional knobbing information.

The knobbing information is computed in the resolver based on the CLIB
algorithm that John posted.