:zap: reduce resources
This version uses only one socket to talk to the DPM. This change reduced
the CPU load by half. It seemed to also reduce the network traffic (because
all the traffic was encrypted over on link instead of each link doing its
own?)

We need to do more tests with this, but it looks very promising.