:zap: use two maps instead of one
One map is a `String` -> config mapping. It's used to store an account's
last configuration. The other map is an `Int` -> config mapping. This is
used for user-defined configurations.

Splitting the maps removes the linear search for finding account configs
and negates the need to have accounts use name-mangling to make them
stand out.