:construction: optimistic that this accesses the JWT
After a lot of starts and stops, this "simple" solution might allow the
resolvers to access the JWT! I came to this solution by committing WIP
to the branch from *acsys-proxy*, cloning it on my laptop, and using
ChatGPT and VSCode's Copilot to explore the various APIs rapidly. Much
nicer experience than what I was doing in *acsys-proxy*.

But now I'm committing this further WIP so I can build and run it on the
real hardware.