:sparkles: plot config uses JWT
The user config resolvers are using the JWT to find the account
information. RIGHT NOW IT'S NOT VALIDATING THE JWT! I wanted John's team
to be able to save and restore config for different users. A future
commit will validate the JWT.

Right now this code needs testing because I don't know whether our JWT
has the account in the field I'm using, so this is still alpha-quality
code.