:memo: fix documentation
When I switch to using the `#[doc = "..."]` form, I was using the
multiline form for strings which are supposed to remove leading space.
But, because we had a blank line, it didn't remove the leading space for
the following paragraphs. The Markdown formatter took that to mean it
was a preformatted block which didn't render well in the web page.

Now the text is wrapping properly.