:zap: optimize the return of one-shot data
The function returns a `Vec<DataReply>`. I originally kept track of
entered elements by using `Vec<Option<DataReply>>` and a set of indices.
This required translating the vector when returning it.

But since we're already keeping track of the entries using the set,
there's no need for the `Option<>`. Now we return the vector we've been
using to accumulate the results.