:recycle: create a new stream type
This rewrites the `.chain()` functionality used by the archived-to-live
data switch-over. The problem with the `.chain()` method is that the
live stream would be polled until *all* the data from the archive stream
was processed. For large historical requests, the live stream producer
(i.e. DPM) might get a timeout error since we wouldn't be pulling from
its gRPC stream.

This creates a new stream that acts like `.chain()` but pulls data from
the live stream and stores it locally. When the archive stream ends,
it'll first send the buffered data before sending new, live data.