:bug: fix a switchover regression
There were a few cases where we were buffering live data forever. This
was due to the archive stream not actually ending when it doesn't have
further information. This commit recognizes when it ends and switches on
the live data. There's also a flushing of the buffered data if both
streams end.