:art: remove use of `.unwrap()`
Even though it was safe to use `.unwrap()`, it was a "code smell". It
was safe because the flag `need_archived` would be true if the start
time existed and false otherwise. But having that decision in two places
(the flag and the actual data) wasn't right.

This commit fixes it by making the start time only `Some()` when it's
before `now`. Then the test for `Some()` is always correct and can't get
out of sync with a flag.