:adhesive_bandage: DPM deficiency work-around
DPM's support for live data vs archived data -- along with its support
for scalars and arrays -- has inconsistent requires for the requests.
Until it gets cleaned-up, we massage the request into what DPM wants to
see.

In this case, the GraphQL client is asking for archived data for an
array device. The client will add the `[]` notation to signify this.
DPM, however, doesn't want to see any subscripts characters, so we
remove it.